requirements.txt

PyGithub==2.4.0
typer==0.12.3
pyyaml==6.0.2
python-dotenv==1.0.1
httpx==0.27.0
